<Feature>[Mermaid Plugin]: <Update link generation>
[
* Include link name
* Fix SetupClass;  Cannot raise an exception or tests do not run
* Fix cleanup so only attempts deletion of files;  This for the case when we want to run tests singly ]

[https://github.com/hasii2011/pyutplugins/issues/49]

+++ tests/pyutplugins/ioplugins/mermaid/TestMermaidWriter.py
@@ -46,13 +46,13 @@ class TestMermaidWriter(TestBase):
         TestBase.setUpLogging()
         TestMermaidWriter.clsLogger = getLogger(__name__)
 
-        try:
+        if 'KEEP' in environ:
             keep: str = environ["KEEP"]
             if keep.lower().strip() == 'true':
-                cls.keep: bool = True
+                cls.keep = True
             else:
                 cls.keep = False
-        except KeyError:
+        else:
             cls.clsLogger.info(f'No need to keep data files')
             cls.keep = False
 
@@ -167,7 +167,8 @@ class TestMermaidWriter(TestBase):
     def cleanup(cls, baseFileName: str):
         path: Path = Path(baseFileName)
 
-        path.unlink()
+        if path.exists() is True:
+            path.unlink()
 
 
 def suite() -> TestSuite:
